Dando funcao aos inputs
