+++ frontendHabitPRO/src/app/login/index.js
@@ -1,13 +1,35 @@
 import React from 'react';
-import { View, Text } from 'react-native';
-import  styles from './styles';
+import { View, Text, Image, TouchableOpacity } from 'react-native';
+import styles from './styles';
+import BasicButtom from '../../components/BasicButton';
+import BasicInput from '../../components/BasicInput';
 
-const Login = () => {
+const Welcome = () => {
   return (
     <View style={styles.container}>
-      <Text style={styles.title}>Login</Text>
+      <Image style={styles.logo}
+        source={require('../../assets/images/icon.png')}
+      />
+      <View style={styles.inputContainer}>
+        <BasicInput
+          placeholder="E-mail" />
+        <BasicInput
+          placeholder="Senha" />
+        <TouchableOpacity style={styles.forgotPasswordButtom}>
+          <Text style={styles.forgotPasswordText}>Esqueci a senha</Text>
+        </TouchableOpacity>
+      <View style={styles.buttomContainer}>
+        <BasicButtom
+          title="Entrar"
+          color="#BCE5FA"
+          textColor="#fff"
+          width="265"
+          height="45"
+        />
+      </View>
+      </View>
     </View>
   );
 };
 
-export default Login;
+export default Welcome;
