+++ frontendHabitPRO/src/app/login/styles.js
@@ -1,17 +1,26 @@
-import { StyleSheet } from 'react-native';
+import { StyleSheet } from "react-native";
 
 export default styles = StyleSheet.create({
   container: {
     flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: 20,
-    backgroundColor: '#f5f5f5',
+    alignItems: "center",
+    padding: 15,
+    backgroundColor: "#f5f5f5",
   },
-  title: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginBottom: 20,
-    color: '#333',
+  logo: {
+    width: 312,
+    height: 95,
+    margin: 20,
+    flex: 0.5,
+  },
+  inputContainer: {
+    flex: 0.6,
+  },
+  forgotPasswordButtom: {
+    flexDirection: "row-reverse",
+    marginHorizontal: "10",
+  },
+  forgotPasswordText: {
+    color: "#7D7777",
   },
 });
