+++ frontendHabitPRO/src/app/welcome/styles.js
@@ -1,0 +1,27 @@
+// src/app/welcome/styles.js
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#FFE8EA',
+    alignItems: 'center',
+    justifyContent: 'center',
+    padding: 20,
+  },
+  logo: {
+    fontSize: 32,
+    color: '#A0D2EB',
+    fontFamily: 'Poppins',
+  },
+  logoBold: {
+    color: '#FAD02C',
+    fontWeight: 'bold',
+  },
+  slogan: {
+    color: '#2D2D2D',
+    fontSize: 16,
+    marginVertical: 20,
+    textAlign: 'center',
+  },
+});
