--- frontendHabitPRO/src/app/welcone/styles.js
@@ -1,26 +1,0 @@
-import { StyleSheet } from "react-native";
-
-export default styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: "center",
-    padding: 15,
-    backgroundColor: "#f5f5f5",
-  },
-  logo: {
-    width: 312,
-    height: 95,
-    margin: 20,
-    flex: 0.5,
-  },
-  inputContainer: {
-    flex: 0.6,
-  },
-  forgotPasswordButtom: {
-    flexDirection: "row-reverse",
-    marginHorizontal: "10",
-  },
-  forgotPasswordText: {
-    color: "#7D7777",
-  },
-});
